requirements.txt
streamlit
pandas
matplotlib
xlsxwriter
python-docx
google-generativeai
openpyxl
